Serial imlementation tested

+++ serial/FFNN4.hpp
@@ -13,24 +13,24 @@ class FFNN4
     Layer hiddenLayer1;
     Layer hiddenLayer2;
     Layer outputLayer;
-    std::vector<float> input;
-    std::vector<float> output;
+    std::vector<double> input;
+    std::vector<double> output;
     ActivationFunctionType activation;
 public:
     FFNN4(unsigned int neuronsInput,unsigned int neuronsHidden1,
     	  unsigned int neuronsHidden2,unsigned int neuronsOutput, ActivationFunctionType act = SIGMOID);
     
-    void setInput(const std::vector<float>& in);
+    void setInput(const std::vector<double>& in);
     
-    void setWeightMatrices(const std::vector< std::vector <float> >& w1, const std::vector< std::vector<float> >& w2, const std::vector< std::vector<float> >& w3);
+    void setWeightMatrices(const std::vector< std::vector <double> >& w1, const std::vector< std::vector<double> >& w2, const std::vector< std::vector<double> >& w3);
     
     // Get inputs from input field and store result in output field.
     void calculateOutput(void);
     
     // Get inputs from in parameter and return results.
-    std::vector<float> calculateOutput(const std::vector<float>& in);
+    std::vector<double> calculateOutput(const std::vector<double>& in);
     
-    std::vector<float> getOutput(void) const;  
+    std::vector<double> getOutput(void) const;  
 };
 
 #endif // FFNN4_HPP
